+++ Where_Is_My_Wife/Assets/_Main/Assets/Input/PlayerInputActions.inputactions
@@ -165,7 +165,7 @@
                 {
                     "name": "",
                     "id": "d69615cd-eecb-4b7b-a0bc-08e293e1dccb",
-                    "path": "<Keyboard>/leftShift",
+                    "path": "<Keyboard>/j",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
@@ -187,7 +187,7 @@
                 {
                     "name": "",
                     "id": "34d7c13f-da55-4a84-af6a-7d2557850272",
-                    "path": "<Keyboard>/leftCtrl",
+                    "path": "<Keyboard>/k",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
@@ -198,7 +198,7 @@
                 {
                     "name": "",
                     "id": "c0ffceba-40af-41cf-81ad-8fb54330ee5f",
-                    "path": "<Keyboard>/k",
+                    "path": "<Keyboard>/l",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
