--- Where_Is_My_Wife/Assets/_Main/Scripts/UI/ControllerTypeImage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d6e2367a29586f439d853fb1890f4aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
