+++ Where_Is_My_Wife/ProjectSettings/TagManager.asset
@@ -62,3 +62,6 @@ TagManager:
   - name: Tire
     uniqueID: 3881497285
     locked: 0
+  - name: UI
+    uniqueID: 3508742867
+    locked: 0
